Refactor ls::find_top_3_insertions.
