+++ src/algorithms/local_search/swap_star_utils.h
@@ -12,6 +12,7 @@ All rights reserved (see LICENSE).
 
 #include <algorithm>
 
+#include "algorithms/local_search/top_insertions.h"
 #include "structures/typedefs.h"
 #include "structures/vroom/input/input.h"
 #include "structures/vroom/solution_state.h"
@@ -22,48 +23,6 @@ All rights reserved (see LICENSE).
 // additional constraint checks (back-hauls and time windows).
 
 namespace vroom::ls {
-
-struct InsertionOption {
-  Eval cost;
-  Index rank;
-};
-
-using ThreeInsertions = std::array<InsertionOption, 3>;
-
-constexpr InsertionOption no_insert = {NO_EVAL, 0};
-constexpr ThreeInsertions
-  empty_three_insertions({no_insert, no_insert, no_insert});
-
-template <class Route>
-ThreeInsertions find_top_3_insertions(const Input& input,
-                                      Index j,
-                                      const Route& r) {
-  const auto& v = input.vehicles[r.vehicle_rank];
-
-  auto best_insertions = empty_three_insertions;
-
-  for (Index rank = 0; rank <= r.route.size(); ++rank) {
-    const InsertionOption current_insert =
-      {utils::addition_cost(input, j, v, r.route, rank), rank};
-
-    if (current_insert.cost < best_insertions[2].cost) {
-      if (current_insert.cost < best_insertions[1].cost) {
-        if (current_insert.cost < best_insertions[0].cost) {
-          best_insertions[2] = best_insertions[1];
-          best_insertions[1] = best_insertions[0];
-          best_insertions[0] = current_insert;
-        } else {
-          best_insertions[2] = best_insertions[1];
-          best_insertions[1] = current_insert;
-        }
-      } else {
-        best_insertions[2] = current_insert;
-      }
-    }
-  }
-
-  return best_insertions;
-}
 
 struct SwapChoice {
   Eval gain;
